+++ frontend/app/component/form/data_entry/PollingStationFormNavigation.tsx
@@ -1,16 +1,15 @@
 import * as React from "react";
 import { BlockerFunction, useBlocker, useNavigate } from "react-router";
 
-import { AbortDataEntryModal } from "app/module/data_entry";
-
 import { Election, PollingStationResults } from "@kiesraad/api";
 import { t, tx } from "@kiesraad/i18n";
 import { Button, Modal } from "@kiesraad/ui";
 
-import { ErrorModal } from "../error";
-import { AnyFormReference, FormSectionID, FormState } from "../form/data_entry/PollingStationFormController";
-import { currentFormHasChanges } from "../form/data_entry/pollingStationUtils";
-import { usePollingStationFormController } from "../form/data_entry/usePollingStationFormController";
+import { AbortDataEntryModal } from "../../../module/data_entry";
+import { ErrorModal } from "../../error";
+import { AnyFormReference, FormSectionID, FormState } from "./PollingStationFormController";
+import { currentFormHasChanges } from "./pollingStationUtils";
+import { usePollingStationFormController } from "./usePollingStationFormController";
 import { getUrlForFormSectionID } from "./utils";
 
 export interface PollingStationFormNavigationProps {
